# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.BoardSummaryResponse:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      done_cards:
        type: integer
      icon:
        type: string
      id:
        type: string
      last_accessed_at:
        type: string
      members:
        items:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MemberSummary'
        type: array
      title:
        type: string
      total_cards:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse:
    properties:
      acceptance_criteria:
        type: string
      assignee_id:
        type: string
      assignee_name:
        type: string
      column_id:
        type: string
      completed_at:
        type: string
      completed_subtasks:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      due_date:
        type: string
      estimated_hours:
        type: number
      id:
        type: string
      implementation_note:
        type: string
      is_done:
        type: boolean
      position:
        type: number
      priority:
        type: string
      tags:
        items:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.TagResponse'
        type: array
      title:
        type: string
      total_subtasks:
        type: integer
    type: object
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.ColumnResponse:
    properties:
      cards:
        items:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse'
        type: array
      category:
        type: string
      color:
        type: string
      id:
        type: string
      position:
        type: number
      title:
        type: string
    type: object
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CreateBoardRequest:
    properties:
      color:
        type: string
      description:
        description: |-
          Appearance is optional at create time. Omit/null = column default
          (description '', color '#1E40AF', icon 'board'). Same validation as the
          PATCH path so a board can be born with its chosen identity in one request.
        maxLength: 160
        type: string
      icon:
        enum:
        - board
        - rocket
        - target
        - bolt
        - bug
        type: string
      title:
        maxLength: 120
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MemberSummary:
    properties:
      full_name:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.StashedBoardDTO:
    properties:
      color:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      stashed_at:
        type: string
      title:
        type: string
    type: object
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.TagResponse:
    properties:
      board_id:
        type: string
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UpdateBoardRequest:
    properties:
      budget:
        minimum: 0
        type: number
      color:
        type: string
      description:
        description: |-
          Appearance. Description "" is a valid clear (column is NOT NULL DEFAULT '');
          color must be a hex string; icon is one of the known glyph keys mirrored
          on the client (lib/boardAppearance). omit/null = no change (COALESCE-style
          read-modify-write in the service).
        maxLength: 160
        type: string
      icon:
        enum:
        - board
        - rocket
        - target
        - bolt
        - bug
        type: string
      title:
        maxLength: 120
        minLength: 1
        type: string
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Turtask
  description: |-
    Multi-board Kanban + task management with realtime sync.
    Auth uses an HttpOnly `auth_token` cookie issued by /api/auth/login or /api/auth/oauth.
  title: Turtask API
  version: "1.0"
paths:
...
  /api/boards:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.BoardSummaryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List my boards
      tags:
      - boards
    post:
      consumes:
      - application/json
      parameters:
      - description: Board title
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CreateBoardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: id
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create board
      tags:
      - boards
  /api/boards/{boardID}:
    delete:
      parameters:
      - description: Board UUID
        in: path
        name: boardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: not the owner
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Stash board
      tags:
      - boards
    get:
      parameters:
      - description: Board UUID
        in: path
        name: boardID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.ColumnResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: not a member or board not found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get board contents
      tags:
      - boards
    patch:
      consumes:
      - application/json
      parameters:
      - description: Board UUID
        in: path
        name: boardID
        required: true
        type: string
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UpdateBoardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.BoardSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "403":
          description: not a manager
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update board
      tags:
      - boards
...
  /api/stash:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.StashedBoardDTO'
            type: array
      security:
      - CookieAuth: []
      summary: List stashed boards
      tags:
      - stash
  /api/stash/{boardID}:
    delete:
      parameters:
      - description: Board UUID
        in: path
        name: boardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Permanently delete stashed board
      tags:
      - stash
  /api/stash/{boardID}/restore:
    patch:
      parameters:
      - description: Board UUID
        in: path
        name: boardID
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - CookieAuth: []
      summary: Restore stashed board
      tags:
      - stash
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"